refactor: small refactoring
