+++ package.json
@@ -41,7 +41,13 @@
             ]
         }
     },
+    "config": {
+        "commitizen": {
+            "path": "cz-conventional-changelog"
+        }
+    },
     "scripts": {
+        "commit": "git-cz",
         "test": "nyc mocha 'test/unit/**/*.spec.ts' 'test/e2e/**/*.spec.ts'",
         "test:unit": "nyc mocha 'test/unit/**/*.spec.ts'",
         "test:e2e": "nyc mocha 'test/e2e/**/*.spec.ts'",
@@ -55,6 +61,7 @@
         "remark": "remark README.md CHANGELOG.md CONTRIBUTING.md CODE_OF_CONDUCT.md .github/ -o -f -q ",
         "pre-commit": "git add . && npm run format:staged && npm run remark && npm run lint && npm run test && npm run build",
         "build": "rimraf dist && tsc",
+        "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
         "prepublishOnly": "npm run build",
         "preversion": "npm run pre-commit",
         "version": "npm run changelog",
@@ -80,7 +87,9 @@
         "axios-mock-adapter": "1.16.0",
         "chai": "4.2.0",
         "chai-as-promised": "7.1.1",
+        "commitizen": "3.0.7",
         "coveralls": "3.0.3",
+        "cz-conventional-changelog": "2.1.0",
         "dotenv": "6.2.0",
         "husky": "1.3.1",
         "lint-staged": "8.1.5",
