+++ src/exception/validation.exception.ts
@@ -3,7 +3,7 @@ import { GeocoderException } from './geocoder.exception';
 
 export class ValidationException extends GeocoderException {
     constructor(private readonly errors: ValidationError[]) {
-        super('Validation error');
+        super('Validation Failed.');
     }
 
     getValidationErrors(): ValidationError[] {
